requirements.txt
yagmail
pyotp
selenium-recaptcha-solver
